refactor: 용어 관련 Command 서비스를 관리자 용어 관련 서비스로 변경
